LayoutTests/http/tests/webxr/webxr-third-party-iframe-makexrcompatible-denied-by-insufficient-feature-policy.html

<!DOCTYPE html>
<html>
<body>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that third-party iframes will be denied from calling makeXRCompatible() without xr-spatial-tracking feature policy.");
jsTestIsAsync = true;

onmessage = (_e) => {
    e = _e;
    if (e.data.makeXRCompatible != undefined) {
        shouldBeFalse("e.data.makeXRCompatible");
        finishJSTest();
    }
};
</script>
<iframe src="http://localhost:8000/webxr/resources/webxr-makexrcompatible-test.html"></iframe>
</body>
</html>
